multiple cols to map - make them into a proper list
