+++ etl2/app/tables/clients/persons.py
@@ -2,6 +2,9 @@ from mapping.mapping import Mapping
 from transformations import transformations_from_mapping
 import pandas as pd
 
+from transformations.generate_source_query_new import (
+    generate_select_string_from_mapping,
+)
 
 definition = {
     "sheet_name": "persons (Client)",
@@ -13,10 +16,15 @@ definition = {
 def insert_persons_clients(config, etl2_db):
 
     mapping_from_excel = Mapping(
-        excel_doc=config.mapping_document, table_definitions=definition
+        excel_doc=config.mapping_document, sheet_name=definition["sheet_name"]
     )
     mapping_dict = mapping_from_excel.mapping_definitions()
-    source_data_query = mapping_from_excel.generate_select_string_from_mapping()
+
+    source_data_query = generate_select_string_from_mapping(
+        mapping=mapping_dict,
+        source_table_name=definition["source_table_name"],
+        db_schema=config.etl1_schema,
+    )
 
     source_data_df = pd.read_sql_query(
         sql=source_data_query, con=config.connection_string
